+++ src/main/java/ferdi/networkbuilder/services/agentcreationdata/AgentCreationDataService.java
@@ -7,5 +7,5 @@ import java.util.List;
 import java.util.Map;
 
 public interface AgentCreationDataService {
-    Map<Long, List<List<List<AgentCreationData>>>> buildAgentCreationData(Map<Long, GeographicHouseholdMetaData> metaData, int maxAgents);
+    Map<Integer, List<List<List<AgentCreationData>>>> buildAgentCreationData(Map<Long, GeographicHouseholdMetaData> metaData, int maxAgents);
 }
